requirements.txt
openai==1.3.0
langchain==0.0.350
langchain-openai==0.0.2
yfinance==0.2.18
duckduckgo-search==4.1.1
python-dotenv==1.0.0
phoenix-otel==0.1.0
tiktoken==0.5.1
textblob==0.17.1 